move keyframes; use correct CTP remainder calculation; refactor midi_clip_get_events to use timeline positions (necessary for correct handling of overlapping clips)

+++ src/tempo.h
@@ -78,7 +78,8 @@ typedef struct click_track_pos {
     int measure;
     int beat;
     int subdiv;
-    int32_t remainder;
+    double remainder; /* as fraction of subdiv len */
+    /* int32_t remainder; */
 } ClickTrackPos;
